+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - New #359: Add `GridView::keepColumnAttributesInEmptyCell()` to keep column body cell attributes on empty cells (@vjik)
 - New #360: Add `GridView::filterRowAttributes()`, and `filterAttributes` and `filterClass` parameters to `DataColumn`
   to set HTML attributes and a CSS class for filter cells (@vjik)
+- New #362: Add `GridView::captionAttributes()` method and `$attributes` parameter to `GridView::caption()` to set
+  HTML attributes for the `caption` tag (@vjik)
 - Enh #358: Make dependency container in `GridView` constructor optional (@vjik)
 
 ## 1.2.0 September 02, 2026